+++ src/common/components/Tabs/index.js
@@ -1,28 +1,38 @@
 // @flow
 
-import Container from 'common/components/Container';
-import Tab from './Tab';
-import styles from './styles.less';
 import findIndex from 'lodash/findIndex';
+
+import Container from 'common/components/Container';
+
 import Head from 'common/components/Head';
+import styles from './styles.less';
+import Tab from './Tab';
 
 import type { Tab$Props } from './Tab';
 
 type TabInput = Tab$Props & {
   content: any,
   ignoreTitle?: boolean,
+  description?: string,
+};
+
+type TabsProps = {
+  tabLayout?: any,
+  tabs: Array<TabInput>,
+  titleSuffix: string,
 };
 
 const zeroIndex = (index) => (index < 0 ? 0 : index);
 
-const Tabs = ({ tabs, titleSuffix }: { tabs: Array<TabInput>, titleSuffix: string }) => {
+const Tabs = ({ tabs, titleSuffix, tabLayout }: TabsProps) => {
   const { pathname } = window.location;
   const selected = findIndex(tabs, (tab) => tab.to === pathname);
-  const { content, name, ignoreTitle } = tabs[zeroIndex(selected)];
+  const { content, name, ignoreTitle, description } = tabs[zeroIndex(selected)];
+  const Layout = tabLayout;
 
   return (
     <div className={styles.root}>
-      {ignoreTitle || <Head title={`${name} | ${titleSuffix}`} />}
+      {ignoreTitle || <Head title={`${name} | ${titleSuffix}`} description={description} />}
 
       <div className={styles.tabsBg}>
         <Container className={styles.tabsContainer}>
@@ -40,7 +50,7 @@ const Tabs = ({ tabs, titleSuffix }: { tabs: Array<TabInput>, titleSuffix: strin
         </Container>
       </div>
 
-      {content}
+      {Layout ? <Layout>{content}</Layout> : content}
     </div>
   );
 };